WUT-4-UT-vacations

+++ src/main/java/ma/yc/service/impl/VacationServiceImpl.java
@@ -2,14 +2,18 @@ package ma.yc.service.impl;
 
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
+import jakarta.validation.ConstraintViolation;
+import jakarta.validation.Validator;
 import ma.yc.entity.Employee;
 import ma.yc.entity.Vacation;
 import ma.yc.exception.EntityNotFoundException;
+import ma.yc.exception.InvalidedRequestException;
 import ma.yc.repository.EmployeeRepository;
 import ma.yc.repository.VacationRepository;
 import ma.yc.service.VacationService;
 
 import java.util.List;
+import java.util.Set;
 
 import static ma.yc.util.Dates.getDaysLeft;
 
@@ -18,16 +22,19 @@ public class VacationServiceImpl implements VacationService {
 
     private VacationRepository repository;
     private EmployeeRepository employeeRepository;
+    private final Validator validator;
+
 
     @Inject
-    public VacationServiceImpl ( VacationRepository repository, EmployeeRepository employeeRepository ) {
+    public VacationServiceImpl ( VacationRepository repository, EmployeeRepository employeeRepository, Validator validator ) {
         this.repository = repository;
         this.employeeRepository = employeeRepository;
+        this.validator = validator;
     }
 
     @Override
     public boolean create ( Vacation vacation ) {
-
+        validate(vacation);
         Employee employee = vacation.getEmployee();
         int sold = employee.getSoldVacation();
 
@@ -47,11 +54,15 @@ public class VacationServiceImpl implements VacationService {
 
     @Override
     public boolean update ( Vacation vacation ) {
+        validate(vacation);
         return repository.update(vacation);
     }
 
     @Override
     public boolean delete ( Vacation vacation ) {
+        if (repository.findById(vacation.getId()).isEmpty()) {
+            throw new EntityNotFoundException("vacation", vacation.getId());
+        }
         return repository.delete(vacation);
     }
 
@@ -74,6 +85,14 @@ public class VacationServiceImpl implements VacationService {
     public boolean setStatusAccepted ( Vacation vacation ) {
         return repository.setStatusAccepted(vacation);
 
+    }
+
+    private void validate ( Vacation vacation ) {
+        Set<ConstraintViolation<Vacation>> violations = validator.validate(vacation);
+        if (!violations.isEmpty()) {
+            violations.forEach(c -> System.out.println(c.getPropertyPath().toString() + " -> " + c.getMessage()));
+            throw new InvalidedRequestException("error vacation validation");
+        }
     }
 
 }
